requirements.txt
accelerate==1.5.2
appnope==0.1.4
asttokens==3.0.0
bleach==6.2.0
bokeh==3.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.4.1
datashader==0.17.0
debugpy==1.8.12
decorator==5.2.0
enum34==1.1.10
executing==2.2.0
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
GephiStreamer==2.0.3
hdbscan==0.8.40
holoviews==1.20.1
huggingface-hub==0.29.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
multipledispatch==1.0.0
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
panel==1.6.1
param==2.2.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyct==0.5.0
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.2.1
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.2.18
tokenizers==0.21.0
toolz==1.0.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
umap-learn==0.5.7
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
ws4py==0.6.0
xarray==2025.1.2
xyzservices==2025.1.0
